Repo cleanup (#57)
Asked AI to look around the repo for CSS and components that we could
get rid of, and I think this is everything. Looks right to me at least.
I won't merge yet though, we should probably click around this a decent
amount to make sure it doesn't blow up styling.